Improve theme settings usability from code
