+++ autoload.php
@@ -9,7 +9,6 @@ function WPChildThemeBoilerplateAutoload($class) {
 
     $len = strlen($prefix);
 
-    // Only load classes within our namespace
     if (strncmp($prefix, $class, $len) !== 0) {
         return;
     }
@@ -17,13 +16,12 @@ function WPChildThemeBoilerplateAutoload($class) {
     $relative_class = substr($class, $len);
     $relative_path  = str_replace('\\', '/', $relative_class) . '.php';
 
-    // Loop through each base directory and require if found
     foreach ($baseDirs as $base_dir) {
         $file = $base_dir . $relative_path;
 
         if (file_exists($file)) {
             require $file;
-            return; // Stop after finding the class
+            return;
         }
     }
 }
