--- inc/SettingsInterface.php
@@ -1,7 +1,0 @@
-<?php
-namespace WPChildThemeBoilerplate;
-
-interface SettingsInterface {
-    public function enqueueScripts();
-    public function enqueueStyles();
-}
