+++ package.json
@@ -7,13 +7,17 @@
     "lib": "lib"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "gulp": "gulp",
+    "build-css": "gulp sass",
+    "minify-css": "gulp minify",
+    "default": "gulp"
   },
   "author": "",
   "license": "ISC",
-  "devDependencies": {
-    "gulp": "^5.0.0",
+  "dependencies": {
+    "gulp": "^5.0.1",
     "gulp-clean-css": "^4.3.0",
+    "gulp-cli": "^3.1.0",
     "gulp-rename": "^2.0.0",
     "gulp-sass": "^6.0.0",
     "gulp-uglify": "^3.0.2",
